order_service/requirements.txt
fastapi
uvicorn
requests
pika
grpcio
grpcio-tools
sqlalchemy
psycopg2-binary